set rollbackableUsage_ 64bits Signed-off-by: A7355007 <xuhuajun3@h-partners.com>

+++ surface/include/buffer_queue.h
@@ -379,8 +379,8 @@ private:
     int32_t defaultWidth_ = 0;
     int32_t defaultHeight_ = 0;
     uint64_t defaultUsage_ = 0;
-    uint32_t rollbackableUsage_ = BUFFER_USAGE_AUXILLARY_BUFFER0;
-    uint32_t rollbackUsage_ = 0;
+    uint64_t rollbackableUsage_ = BUFFER_USAGE_AUXILLARY_BUFFER0;
+    uint64_t rollbackUsage_ = 0;
     uint32_t bufferQueueSize_ = SURFACE_DEFAULT_QUEUE_SIZE;
     ScalingMode scalingMode_ = ScalingMode::SCALING_MODE_SCALE_TO_WINDOW;
     GraphicTransformType transform_ = GraphicTransformType::GRAPHIC_ROTATE_NONE;
